VHM:Update cpu id type as uint16_t for struct acrn_create_vcpu
Update the cpu id type as uint16_t for struct acrn_create_vcpu in
the VHM driver, this structure is for data transfering between
the hypervisor and device modle in SOS.

Tracked-On:

+++ include/linux/vhm/acrn_common.h
@@ -178,10 +178,10 @@ struct acrn_create_vm {
  */
 struct acrn_create_vcpu {
 	/** the virtual CPU ID for the VCPU created */
-	uint32_t vcpu_id;
+	uint16_t vcpu_id;
 
 	/** the physical CPU ID for the VCPU created */
-	uint32_t pcpu_id;
+	uint16_t pcpu_id;
 } __attribute__((aligned(8)));
 
 /**
